refactor(api): extract audit handlers from system.rs
Move audit_recent and audit_verify from routes/system.rs into the
existing routes/audit.rs module, alongside the related audit_query
and audit_export endpoints. Public paths (/api/audit/recent,
/api/audit/verify) are unchanged.

Refs #3749 (7-of-N)
